Assignment M4 draft
I added lots of attributes. Im not sure if we need methods but feel free to add ig. My syntax and logic are correct as far as i can tell. I know that the super constructor accesses the parent constructor so i put a for loop that copies the tag array from the implemented array. The tag array could come through the scanner in the runner so i didnt think we needed that here. Lemme know lads